services/booking-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
motor==3.3.2
pymongo==4.6.0
strawberry-graphql[fastapi]==0.217.1
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1
httpx==0.25.2
aio-pika==9.3.1
python-dotenv==1.0.0
python-multipart==0.0.6
bcrypt==4.1.1
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1